Version 3.54 is here!
Now a user my make as many question banks as he/she require, can host as many exams as he/she wants, and can give a particular test once.
Many changes are made into the database structure, as it was in the beginning.
Log in and sign are also improved.
Questions will be saved in a different database 'questionbank', and all the other tables in the database 'quizzer'. This is to maintain size of both the databases.
User details is a completely different table now which is mere a informative table now, with a 'handle' which is the only and most important part of the "Quizzer".