Version 4.00 is launched!!
The user is now able to edit and even delete the questions in any of his question banks. He can alter every information of the question if he wishes to or he can even delete the questions.
Next update will be having notification for saved option during exam i.e. in the displayQuestions.php , and feature to delete a question bank, also to close the exam and print scoreboard.